# Add SetHoldings for a collection of targets